+++ src/actions/index.js
@@ -1,0 +1,2 @@
+export * from "./actions"
+export * from "./loyaltyPoolActions"
